Convert .mk files to .bp files
After initial conversion the build failed because the ike target
depends on the bouncycastle target which is only for use by core
libraries. The checks to prevent that only work on .bp files which is
why that was not detected previously.

The solution was simply to remove the bouncycastle target as that
dependency was actually unnecessary because the APIs required by ike
are provided as part of the core.platform.api.stubs which is added to
the compilation classpath by default for any target that does not
specify an SDK version.

Bug: 73635411
Test: atest FrameworksIkeTests
Change-Id: I34efd6abfe3aaea0f5e175fb4a7f1a4617e82acc