Add support for locally initiated IKE deletion
This change adds support for locally initiating a delete IKE request. In
addition, the following was modified:
- IkeCipher, IkeMacIntegrity and IkeMacPrf are no longer final, to
allow for mocking. They are for internal use only, so final is not
strictly necessary.
- Added a method to the IkeSessionStateMachine class to initialize the
StateMachine in the IDLE state, allowing for testing without going
through the IKE_INIT/IKE_AUTH flow every single time.

Bug: 122737998
Test: New tests added, FrameworksIkeTests passing
Change-Id: Ie0d2d367f907efeeae0fcc33262fe3fdbb4a1302