Refactor EapSimTypeData decoding.
The current functionality for decoding EapSimTypeData objects makes it
difficult to unit test the EapSimMethodStateMachine. This refactor moves
the decoding to a EapSimTypeDataDecoder (defined as a static class in
EapSimTypeData). This class can be provided in a @VisibleForTesting
EapSimMethodStateMachine constructor, meaning mocks can be easily
injected for testing functionality with specific EapSimTypeData
instances.

Bug: 135609658
Test: atest EapSimTypeDataTest
Change-Id: I9998ee1c869a8e659656cbd22fa7ac2bd64bc8b3