Persist EAP-SIM Identity in StartState.
When the identity is requested by the authentication server, it should
be made part of the state in the StartState. The Identity value is
necessary in the ChallengeState for calculating the Master Key (MK),
from which the K_aut, Master Session Key (MSK), and Extended Master
Session Key (EMSK) are derived. By persisting the identity value from
the StartState, it can be easily transferred to the ChallengeState via
the constructor.

Bug: 136280277
Test: atest EapSimStartStateTest
Change-Id: Idcc3b6d2a4f701974db4f3e1c79109b38602a425