Add integration testing for EAP-SIM.
EAP-SIM currently has one integration test that checks the happy path.
Tests are added to verify that the library correctly NAKs an
unsupported, unconfigured EAP method. The library subsequently switches
to the configured EAP method. Tests are also added to check that
EAP-Notifications can be received at every step of the EAP-SIM process.

Bug: 139367931
Test: atest EapSimTest
Change-Id: I260276161b65f3a394aa768f8af4dcd7e9e0114b