Refactor EAP-SIM attributes to incorporate EAP-AKA.
EAP-SIM and EAP-AKA share most of the same attribute definitions, and
are defined side-by-side by the IANA.
- EapSimAttribute is refactored to be EapSimAkaAttribute, where all
EAP-SIM and EAP-AKA attributes (not yet defined) will be defined.
- Parsing logic through EapSimAttributeFactory is restructured to rely
on EapUiccAttributeFactory for decoding any attributes used by both
EAP-SIM and EAP-AKA.

The EAP library is also refactored to allow for cleaner, alternate EAP
method implementations. EAP-SIM and EAP-AKA are put together in several
sub-packages (simaka).

Bug: 139381128
Test: atest com.android.ike.eap
Change-Id: Id967113e07c125788600daa4b80c45c4da1e1a7f