Implement UICC challenge processing for EAP-AKA.
EAP-AKA needs to run the received RAND and AUTN values through the
authentication procedure. The response value needs to be parsed to
verify that it's either a valid challenge response, a synchronization
failure, or an unknown response type.

Tags are embedded inside TelephonyManager's base64 response value to
differentiate between a successful authentication and a synchronization
failure. Specifically, 0xDB signals a success and 0xDC signals a
failure. These are the same values used in Telephony and Wifi for
validating 3G authentication.

Bug: 133879622
Test: added tests to EapAkaChallengeStateTest.
Test: atest FrameworksIkeTests
Change-Id: I54ca8f5cb352058b88e1d23d8fe4831970f9ed91