Implement MAC check and response for EAP-AKA.
After processing the RAND request, EAP-AKA needs to verify the received
MAC and generate a response to the authentication server.

If the received MAC doesn't match the MAC computed with keys derived
from the authentication procedure, a Client-Error response is sent to
the server. Once the received MAC is received, a MAC is computed over
the response message (including the AT_RES attribute) and the
response-packet is returned.

Bug: 133879622
Test: added happy-path test to EapAkaChallengeStateTest.
Test: atest FrameworksIkeTests
Change-Id: I478d870ad6cf4c40a880b0006524e9f90df8f415