Implement ValidateAuthenticatorState for EAP MSCHAPv2.
The ValidateAuthenticatorState handles EAP MSCHAPv2 SuccessRequest and
FailureRequest messages. If a SuccessRequest contains a valid
Authenticator Response, the state machine transitions to the
AwaitingEapSuccessState and a SuccessResponse is returned. If a
SuccessRequest contains an invalid Authenticator Response, the state
machine transitions to the FinalState and an EapFailure object is
returned. If a FailureRequest is received, the state machine transitions
to the AwaitingEapFailureState and a FailureResponse is returned.

Bug: 140322003
Test: added EapMsChapV2ValidateAuthenticatorStateTest.
Test: atest FrameworksIkeTests
Change-Id: I4960d1709af99930237dedf04458d4092b3efa9b