Implement AwaitingEapFailureState for EAP MSCHAPv2.
EAP MSCHAPv2 uses theAwaitingEapFailureState for message-handling after
receiving - and responding to - EAP MSCHAPv2 FailureRequest messages. In
this state, EAP-Failure messages are expected and after reception, the
state machine transitions to the FinalState and returns an EapFailure.

If EAP-Success messages are received, they are logged. The state machine
then transitions to the FinalState and returns an EapFailure.

If EAP-Request/Notification messages are received, they are processed
and responded to with EAP-Response/Notification messages.

Any other EAP messages are logged and the appropriate EapError is
returned.

Bug: 141483998
Test: added EapMsChapV2AwaitingEapFailureStateTest.
Test: atest FrameworksIkeTests
Change-Id: Ieba8ef5a0709d61b01d70fb1db68d59935659e89