Check for bidding down attack in EAP-AKA.
EAP-AKA' specifies the use of AT_BIDDING in EAP-AKA as a means to
prevent bidding down attacks from EAP-AKA' to EAP-AKA (RFC 5448#4). To
do so, the EapAkaMethodStateMachine must check for this attribute and
whether EAP-AKA' is supported after successfully authenticating the
server. If the server specifies that EAP-AKA' is supported, an
Authentication-Reject response is sent to the server.

Bug: 142742437
Test: added test case to EapAkaChallengeStateTest.
Test: atest FrameworksIkeTests
Change-Id: I859449c7f537ffe81d7d30c181008430abadbc96