Maintain state for expecting EAP-Failure messages.
EAP methods can enter states where EAP-Failure (and possibly
EAP-Notification) messages are the only valid incoming messages. This
state is not tracked, so a boolean flag is added to
EapMethodStateMachine. This flag is set when states enter an "expecting
error" state. Examples of this case would be after EAP-SIM sends an
EAP-Response/SIM-Client-Error message, or after EAP-AKA sends an
EAP-Response/AKA-Authentication-Reject message.

Bug: 140136397
Test: atest FrameworksIkeTests
Change-Id: I089f66b6eda75bb6d4d88533018c0f1ebe7b883a