Check EAP-SIM for valid challenge before Success.
EAP-SIM should check that it fully processed and responded to a
Challenge Request before allowing EAP-Success messages. This prevents
the case where a malicious server could force an EapError response in
the ChallengeState, then send an EAP-Success message. Without a check
for having processed a valid challenge, the EapSimMethodStateMachine
will provide uninitialized MSK and EMSK values to the client.

Bug: 144027934
Test: atest FrameworksIkeTests
Change-Id: I3dc308d3d03e8db35ee81f6814c9aac48d6ed36a