Use JarJar to test using statically linked copy
Due to IKEv2 being part of the boot classpath, the classloaders that
load it are different from that which loads the tests. As a result, the
package-protected test classes/fields/methods are not accessible to
tests, and causes failures in resolving methods due to visibility.

The solution here is to remap all source and test classes to use a
different package name that would not get overshadowed by the boot
classpath version.

Bug: 143983419
Test: Tests pass with IKE in boot classpath
Change-Id: I769f0ebb0e29af58355d6c2d0cf42e92d0bba800