Platform shouldn't directly link to jars in APEXes
The non-updatable part of the platform shouldn't directly link to the
boot jars in APEXes. Ensure this by

1) setting the visibility property for the boot jars so that they are
not visible to non-APEX modules and
2) setting the apex_available property so that the boot jars are only
built for the corresponding APEXes, but not for others.

Bug: None
Test: m

Exempt-From-Owner-Approval: cherry-pick from internal

Change-Id: I7e17aa6dc1c4efe58d24ee3f29fc5e83230b927e
Merged-In: I7e17aa6dc1c4efe58d24ee3f29fc5e83230b927e