Add ike-internals target as a placeholder
Added an empty ike-internals library that must be added to any target
that compiles ike-srcs. This is broken out as a separate change to
simplify the process of updating those affected targets by avoiding
having to make changes in frameworks/base, frameworks/opt/net/ike and
external/bouncycastle simultaneously in conflicting branches.

Bug: 152328618
Test: m droid
      atest FrameworksIkeTests
Change-Id: I98eea77a70eea6d78d06fbdeaf7997cdcbd2efd4
Exempt-From-Owner-Approval: already had +2 and fixed issue that was raised.