Reschedule rekey in rekey state instead of TempFailureHandler
This commit does not schedule retry in TempFailureHandler anymore
but in the current state. Reason is that some events like rekey
should be scheduled by AlarmManager instead of in the handler.
This commit:
- Adds a method stub #rescheduleRekey in SaRecord
- Schedule rekey retry in rekey state

This CL is a preparation CL before we can switch to
use alarm manager to schedule the initial rekey and rekey retry.

Bug: 149058810
Test: atest FrameworksIkeTests
Change-Id: If4bbf9c889f130308839c4aa57334951a78f60dd
Merged-In: If4bbf9c889f130308839c4aa57334951a78f60dd