Enable api lint and check_last_api for modules
This adds checking of module api compatibility to the individual module
api rules. Until now, this checking has been done via the monolithic
metalava runs which we are aiming to get rid of.

Now is a good time to do this because we can compare them to the just
finalized version 30 API, which we have no diffs with.

Bug: 144149403
Test: m checkapi
Change-Id: Ia74b8f8408244b15f28400dfadbd6a0c292f7953