Fix NullPointException in creating first Child
This commit fixes the NullPointException when creating first
Child is rejected by a error notification. The root cause is
that when receiving a error notification, saPayload returned
by ikeMessage.getPayloadForType will be null and will be
passed to #handleFirstChildExchange.

This commit also stops throwing fatal exception to
IkeSessionStateMachine when there is a missing payload
for create first Child. This will allow IKE AUTH succeed and
let ChildSessionStateMachine to do the check

Bug: 148287674
Test: atest testIkeAuthHandlesFirstChildCreationFail
Test: atest FrameworksIkeTests
Change-Id: I0f130abca20c748969b28cb7ad67de3392a3aa06