Fix NPE in race between child response and remote retransmission
IkeSessionStateMachine will get a NullPointerException when it
receives a retransmitted request of the first remote request
before it has sent back a response. The cause is that the
IkeSessionStateMachine tries to send back the cached response but
the cached response value is null.

This is a tactical fix to ensure correctness.

Bug: 159133972
Test: FrameworksIkeTests, CtsIkeTestCases
Change-Id: I97b7af7cde93b623b24f4dd98c338d19c18437b6