Fix bug in handling retransmitted request
This commit fixes the bug where IkeSessionStateMachine will
send out wrong response when it receives a retransmitted request
before the original request has been processed.

When the remote server initiate a Rekey/Delete Child request, it
is possible that before the Child Session has sent out the
response, the IKE Session receives the retransmitted request.

This CL allows IkeSessionStateMachine to know if the retransmitted
request has been replied or is still being processed

Bug: 159069391
Test: FrameworksIkeTests, CtsIkeTestCases
Change-Id: If5a5778a7f040a4acecc77a48f25482cbd4340cd