Notify remote on killSession().
This change updates IkeSessionStateMachine#killSession to process the
kill command on the Handler thread. The Idle state (as a non-procedure
state) is also updated to notify the remote that the IKE session is
being deleted.

All states will continue the existing behavior for #killSession (closing
all Child SAs and notifying the user callback).

Bug: 150327466
Test: atest FrameworksIkeTests CtsIkeTestCases
Change-Id: Ieb7a297552b2d7c0e89cbe6ba5a0fad809d66e58