Address errorprone presubmit warnings
Got bunch of warnings when running errorprone presubmit on
git_master and made it very hard to locate the real error.

This CL addesses all warnings according to the
errorprone guidance, except 3 warnings for "mocking final
class" because there is non-trivial way to fix that.

The full list of warnings are as followed:
https://android-build.googleplex.com/builds/pending/P15925085/errorprone/latest/view/logs/error.log

Bug: 162373789
Test: FrameworksIkeTests, CtsIkeTestCases
Change-Id: I89167499c36b8c69b18109959c604bb0728a7359