Expose ChildSessionParams as public APIs
As required by API Council in b/148596690, IKE API should be
public because IKE library runs in the client process and it
is not feasible to restrict the code with permissions. In
addition, VCN also requires part of IKE API to be public for
configuring IKE and Child Sessions.

This CL makes System APIs in ChildSessionParams and its two
subclasses public APIs without any change.

This CL also adds @SuppressLint to work around API lint warnings
and fix few typos in documentation

Bug: 151984042
Test: FrameworksIkeTests, CtsIkeTestCases
Change-Id: I06833432d0afa4cca096ab67b2b7fd197e945fbe