Add a getter to return the caller configured Network
Before this CL, if caller does not configure a Network in
IkeSessionParams, the Builder will resolve a default Network
while building an IkeSessionParams, and
IkeSessionParams#getNetwork will return the resolved default
Network. This is misleading because if MOBIKE is enabled,
IKE Session may switch to a different default Network.

This CL adds another getter to return configured Network which
will be exposed as a public API.

This CL also delays the default network resolving process to
IkeStateMachine

Bug: 151984042
Test: FrameworksIkeTests, CtsIkeTestCases
Change-Id: Id3a0f0f7640525fdc8ecf3f7331e23a1abb1cb79