Expose IkeSessionParams.Builder no-arg constructor
This commit deprecates the IkeSessionParams.Builder that requires a Context
object and replace it with a no-arg constructor.

Previously the input Context was used for getting a ConnectivityManager
to resolve a default network. The purpose is to make sure there is
default network during IkeSessionParams creation and allow caller
to retrieve it from IkeSessionParams#getNetwork().

However both the resolving and checking work is unnecessary because
1) Default network will change during IKE Session's lifetime
2) Even if there is an active network during IkeSessionParams creation,
   it may disconnect when IKE Session starts the setup process. That
   is to say, it is meaningless to do this check before setup process.

Thus we want to replace Builder(Context) with Builder()

Followup commit will make getNetwork from @NonNull to @Nullable.

Bug: 163604823
Test: FrameworksIkeTests. CtsIkeTestCases
Change-Id: I8a7b7c80c5e5f892dd4372a579c5926fa126d4fc