Construct IkeSocket in a central place in statemachine
This commit adds two methods: buildIkeSocket,
buildAndSwitchToIkeSocket so that constructing IkeSocket is
centralized.

This commit also swaps the order of updating mLocalAddress and
restarting NAT-T keepalive in onUnderlyingNetworkUpdated to
fix the ClassCastException for casting v6 address to Inet4Address

This commit is a preparation for allowing enforcing IKE Session
to use port 4500.

Bug: 180652705
Bug: 182956310
Test: FrameworksIkeTests, CtsIkeTestCases
Change-Id: I8158ad9670970084b0fd50fbb3a5e224d27e6ce6