Fix bug that getNetwork returns null when Network was set
The root cause is IkeSessionParams.Builder#build() only assigns
value to defaultOrConfiguredNetwork when mConnectivityManager
is not null. In this case if network was set, getNetwork will
return null.

The correct behaviour is even if mConnectivityManager is null, if
setNetwork was set, defaultOrConfiguredNetwork should be assigned
with the configured network.

This CL also replaces setConfiguredNetwork with setNetwork in
tests

Bug: 182415029
Test: atest FrameworksIkeTests
Change-Id: Ic0923cb41e603330ccbf2057f3cbe7d9bcf36c9d