Pull TunNetworkContext to a separate base class
This commit pulls TunNetworkContext from IkeSessionTestBase to
IkeTestNetworkTestBase. Test class that needs to use TunNetworkContext
but does not need to use it in every test method (like
IkeExceptionTest in the followup commit) can inherit
IkeTestNetworkTestBase. Reducing the call to get a test network
can also reduce the test flakiness.

This commit also makes TunNetworkContext Closeable so that it
can be used in a try-with-resource block.

Bug: 183239538
Test: CtsIkeTestCases verfied on S and R platform
Change-Id: Ie8e1d7177113b53724877b7d89aa9653d3349c76