Don't switch IkeSocket if sockets are the same.
This CL updates IkeSessionStateMachine to only switch IkeSockets if the
old and new sockets are different. This fixes a bug where an IKE
Session using IPv4 address and IKE_OPTION_FORCE_PORT_4500 that
encounters a NAT attempts to switch from a UDP Encap socket to a UDP
Encap socket, which caused a use-after-free error.

Bug: 186818479
Test: atest FrameworksIkeTests CtsIkeTestCases
Change-Id: I423d7b7c5a95522c4a8b67d91ffa4f66eb2a44db