Do not send broadcast when running in system uid
This commit uses the callback version of #setExact instead of the
broadcast version. It also uses #setExact instead
of #setExactAndAllowWhileIdle when the current process uid is
system uid and the alarm needs to be fired in doze mode.

The method #setExactAndAllowWhileIdle can only fire an alarm
by sending a broadcast. For any system service that wants to
send broadcast, they need to explictly declare the broadcast
as protected in the manifest, which is not user friendly.

Since system service can fire alarm in doze mode even if the alarm
is set without "AllowWhileIdle", the user inconvenience can be
avoided by changing IKE to use #setExact with a direct callback
instead of using #setExactAndAllowWhileIdle

Bug: 180653278
Test: FrameworksIkeTests (new tests added), CtsIkeTestCases
Test: manually tested alarms fired correctly
Change-Id: I54b7a2bbfa5af7ee487476b87d414607e6100081