Remove wrong payloads from Rekey Delete response
This commit clears mOutboundRespPayloads after sending out a response
for Child exchange to make sure that payloads in the outbound
Rekey Child Create response will not be included in the Rekey Child
Delete response. This fix makes IKE client more spec compliant and
will reduce the failure risk in Rekey Child.

An outbound response for a Child exchange is built based on the
existing mOutboundRespPayloads and additional Child outbound
payloads. This design is for handling a server request for deleting
multiple Child SAs. However, mOutboundRespPayloads is cleared
only in #enterState() and will cause problems in Rekey Child.
Specifically after sending out a Rekey Child response, rekey-create
payloads are still in mOutboundRespPayloads and will be included in
the Rekey Child Delete response. This commit fixes this issue by
clearing mOutboundRespPayloads after sending out a response.

This commit also changes to initialize IkeMessage#ikePayloadList
with a new List instead of the reference of the input List. In this
way changing the input List won't affect the IkeMessage instance.

Bug: 188930071
Test: FrameworksIkeTests (new tests), CtsIkeTestCases
Change-Id: I6564878dfa08ce85c264ad00cf8ea6f597429742