Fix CTS which is broken by S finalization
This commit uses @IgnoreAfter(R) to replace
@SdkSuppress(maxSdkVersion = -1, codeName = "REL") for
tests that need to be skippeed on S/S+ devices.

There are few CTS tests that should be skipped if the current sdk
is larger than R. However, after sc-dev branch has changed S sdk
version to 31 and changed code name to "REL", those tests are not
skipped as expected.

To fix that, this commit uses @IgnoreAfter(R) that will work no
matter S is finalized (e.g. sc-dev) or not (e.g. aosp).

Bug: 191122988
Test: atest CtsIkeTestCases
Test: verified sc-dev test on http://ab/I54700009453517383
Change-Id: I7495cd0b709446227791456f9486dca35e75cdfb