Fix retransmitter to retransmit identical packet
This commit fixes the EncryptedRetransmitter so that retransmitted
packets are identical to the orginal packet. Without this commit
the IKE library is not rfc compliant and an IKE server may ignore
the retransmitted request and cause IKE Session to disconnect.

Bug: 195201458
Test: atest FrameworksIkeTests(new tests), CtsIkeTestCases
Test: manually verified
Original-Change: https://android-review.googlesource.com/1789496
Merged-In: Ieb7cde97bfbabd871a5c344aa5e990272f12e736
Change-Id: Ieb7cde97bfbabd871a5c344aa5e990272f12e736