Support Sending INITIAL CONTACT Notify payload
This commit indicates that the sender supports
INITIAL CONTACT functionality for the IKE Session.

If this option is set for an IKE Session, The
INITIAL CONTACT  message is sent to the peer
to indicate this IKE SA is the only IKE SA
currently active between the authenticated
identities.

Bug: 186179788
Test: atest FrameworksIkeTests(new tests added)
Original-Change: https://android-review.googlesource.com/1864830
Merged-In: Icbff33565aaf3ac9a42e17734a42704e00bb2db8
Change-Id: Icbff33565aaf3ac9a42e17734a42704e00bb2db8