Add a separate flag to configure IKE to Update SA addresses with Rekey
This ensures that the rekey mobility feature is properly set behind
a separate IKE option. Rekey mobility was incorrectly toggled under
the same flag that enables RFC MOBIKE, under the assumption that rekey
mobility had the same properties as RFC MOBIKE, but that assumption has
been proven false, causing critical bugs.

The addition of REKEY_MOBILITY being coupled to RFC MOBIKE was a bug,
and this fixes that for T onwards. In order to ensure the long-term
maintainability of IKE, this feature has been split out to it's own
IKE option, while ensuring backward compatibility for Android
S/S-V2 callers.

Bug: 223680891
Test: atest FrameworksIkeTests (new tests), CtsIkeTestCases
Change-Id: I3dc8ee70ec9f7fafdfa396688ed85409c88d9c32