Catch all exceptions in IkeConnectionController
This commit updates IkeConnectionController to catch all the
exceptions thrown from the calls that are initiated from
NetworkCallbacks and the packet reading poll-loop. Thus these
calls will never crash the process.

Note, the rest of the calls into IkeConnectionController are
all from the IkeSessionStateMachine and there is already a
mechanism to prevent them from crashing the process.

Bug: 219617270
Test: atest FrameworksIkeTests(new tests), CtsIkeTestCases
Change-Id: I3997e21e53118d3ce6055d75f79fef7756543659