Get LinkProperties from onLinkPropertiesChanged callback
This commit changes IkeDefaultNetworkCallback to notify
IkeNetworkUpdater of the network update after onLinkPropertiesChanged
is called, and at the same time pass the LinkProperties to
IkeNetworkUpdater.

Previously IkeConnectionController called getLinkProperties in
onAvailable callback which may result in getting a null value
becuase the return value of the getter might not be the current value.

This commit also updates IkeConnectionController to reject the
setNetwork call if the LinkProperties of that network is null.

Bug: 221379450
Test: atest FrameworksIkeTests(new tests), CtsIkeTestCases
Change-Id: I27f20114dd388e9be42551f295a8b3b5702ac2bd