Enable auto IP family selection
This commit allows IKE Session to automatically select
address families.

In the short term, the auto IP family selection feature will
configure IKE Session to prefer IPv4 so as to improve
the IKE reliability on network that has poor IPv6 support.

In the long term, the auto selection feature will expand
to have more intelligence that allows IKE session to select
the better IP version based on the network environment.

Bug: 241980596
Test: atest FrameworksIkeTests(new tests), CtsIkeTestCases
Change-Id: I6979fca76ce0a73cf6c4f3dbd9702578d47f3877