Skip sending Delete on fatal errors in IKE AUTH
This commit updates the IKE library to be more RFC compliant so that
during IKE AUTH, if IKE session receives error notifications of
Invalid Syntax, Auth Fail, or Unsupported Critical Payload, IKE Session
will not send Delete request anymore. In this case, IKE Session
should be killed without needing to notify the server.

It is safe to not SDK gate this change because the existing behavior
is not RFC compliant and thus there is very low chance for a server
to depend on it. Even if the dependency exists, the only effect is
it will take longer for the server to detect the connection is
terminated and should have very trivial impact to end users.

Bug: 215308946
Test: atest FrameworksIkeTests(new tests), CtsIkeTestCases
Change-Id: I84b9813f1498dba95ba3dc5d896a5e86a1f784c8