Add extra IKE option for automatic on/off keepalive
Define a new IKE option to allow caller to opt-in automatic
on/off keepalive feature.

This commit allows IPsec modules to read the IKE option from
IkeSessionParams to take extra keepalive start options, so
that a VPN client could choose to opt-in to this mechanism and
it will not impact IPsec clients who do not need this mechanism,
e.g. iwlan service.

Bug: 259000745
Test: atest FrameworksIkeTests
Change-Id: I226efd90845e355b6a40e08de9200b2c62fd5590