Add an API to set IP protocol, encap type, and keepalive delay.
This patch adds an API to set the IP protocol, encap type,
and keepalive delay. All these parameters can be passed in
at session setup time, or when specifying a new underlying
network.

If the caller passes in an IP version that cannot be used, the
session fails asynchronously with a callback exception.

Test: in progress
Change-Id: Ib006d2f076baf2095f42c923157e442feb01624c