Allow caller to force MOBIKE with setNetwork
This commit updates IKE to always perform the full process of
MOBIKE when it is requested by the IKE caller with #setNetwork.
Previously, IKE would skip MOBIKE when the network and addresses
did not change.

This allows IKE caller to use MOBIKE to recover the IPsec
connection by updating the addresses of IPsec SAs.

This behavior change only takes effect on Android U and above.

Bug: 261499808
Test: atest FrameworksIkeTests(new tests), CtsIkeTestCases
Change-Id: I2fbebd461cadf7ae15f0770f446a7dcd487a142c